Delay deserializing each task's callable during execution
Covalent's executors currently deserialize the callable in the server
process. We introduce a light wrapper function which deserializes and
invokes the callable and send the wrapper function to the Dask cluster
or remote computer. This has two advantages:

* Since the Covalent dispatcher never touches the deserialized
callable, it doesn't need any of its dependencies.

* When the the remote executor deserializes the wrapper function, the
core callable remains serialized. This allows the wrapper to perform
preparatory steps (such as running shell commands to install deps)
before deserializing the callable.