Refactor Deps classes
This change should allow all Deps to be implemented in a unified way.

Subclasses are now implemented by defining a custom python
function to apply the dependencies, and passing that function and
associated data to the Base Deps constructor.

The Base Deps serializes the function and arguments into
TransportableObjects, and `apply()` simply returns a tuple of
TransportableObjects:

`(serialized_fn, serialized_args, serialized_kwargs)`