Implement `lattice.build_graph_encoded()`
Same interface as `build_graph` except args and kwargs are
assumed to comprise of TransportableObjects.

TransportableObjects which encode fundamental types (determined by
JSON-encodability) are deserialized before invoking the workflow
function. It is assumed that only such arguments are used to determine
control flow in the workflow (e.g. `for i in range(iterations)` where
`iterations` is a lattice-level argument).