Re-dispatching workflows: comparing transport graphs
Add methods for determining which portion of an existing transport
graph can be reused in another transport graph.

The core function `max_cbms()` computes a common subgraph C of two
rooted DAGs A and B with the following properties:

* C is "backward-maximal" in the sense that all paths from the root
that terminate in C are contained in C.
* C is the largest such subgraph

If A and B represent computational graphs, C describes the tasks in A
and B with the same dependencies.