Don't propagate TaskRuntimeError out of `executor.execute()`
When the task exits abnormally in the backend, the implementation of
`run()` is responsible for retrieving the exception traceback,
ensuring it is printed to stderr, and raising a
`TaskRuntimeError`. This will be caught by `execute()` and returned to
`_run_task()` together with any stdout and stderr.

`_run_task()` checks whether `execute()` caught a `TaskRuntimeError`
and sets the task status accordingly. The task stdout and stderr are
captured as long as the error originated from the executor backend and
not from Covalent's interaction with the backend.