Unify `_run_task`'s handling of `execute()`
Both `BaseExecutor` and `AsyncBaseExecutor now expose an async
`_execute()` method. The latter simply awaits the real `execute()`
coroutine, while the former runs the real `execute()` function in a
threadpool.

With this change, `_run_task` handles both `BaseExecutor` and
`AsyncBaseExecutor` in the same manner without having to inspect the
type of the executor. Put differently, the polymorphism is now handled
purely at the executor level.