Improve handling of electron errors during runtime (#1391)
* Introduce `TaskRuntimeError` exception

To be raised by `executor.run()` when an exception occurs in the
backend.

When `_run_task()` catches such exceptions, their description is
written to the node stderr.

* Don't propagate TaskRuntimeError out of `executor.execute()`

When the task exits abnormally in the backend, the implementation of
`run()` is responsible for retrieving the exception traceback,
ensuring it is printed to stderr, and raising a
`TaskRuntimeError`. This will be caught by `execute()` and returned to
`_run_task()` together with any stdout and stderr.

`_run_task()` checks whether `execute()` caught a `TaskRuntimeError`
and sets the task status accordingly. The task stdout and stderr are
captured as long as the error originated from the executor backend and
not from Covalent's interaction with the backend.

* Update Result status if _run_planned_workflow fails

* Upgrade to DB schema V13

Electron.info_filename -> Electron.error_filename

* Update models in ui to use Electron.error_filename

Display both error and stderr in the NodeDrawer's ErrorCard.

* Improve _ExecutorManager.get_executor's exception

Add detail to ValueError

* Add `Result.get_failed_nodes()`

* Enumerate failed nodes in Result.error

* Print stdout in electron ErrorCard

* Print node names in lattice-level error

* Don't write to `Result._error` in failed node handler

Reserve that for `_run_planned_workflow` to summarize all failed
tasks.

* Put `BaseExecutor.teardown()` in finally block

* Remove debug statement

* Allow suppressing tracebacks for exceptions from Covalent

Tracebacks can be enabled by setting `sdk.log_level = "debug"` or by
starting the server in dev mode.

* Assign private file descriptors to each task for stdout and stderr

Each executor gets two properties
- `self.task_stdout: io.StringIO`
- `self.task_stderr: io.StringIO`

which are instantiated at the beginning of `execute()`.

In each executor's `run()`, implementation, the task's stdout and
stderr should be printed to `self.task_stdout` and `self.task_stderr`,
which will be returned to the dispatcher when `execute()` returns.

* Fix local and dask tests

* Add unit tests for `run_workflow()`

* Unify `_run_task`'s handling of `execute()`

Both `BaseExecutor` and `AsyncBaseExecutor now expose an async
`_execute()` method. The latter simply awaits the real `execute()`
coroutine, while the former runs the real `execute()` function in a
threadpool.

With this change, `_run_task` handles both `BaseExecutor` and
`AsyncBaseExecutor` in the same manner without having to inspect the
type of the executor. Put differently, the polymorphism is now handled
purely at the executor level.

* Improve test coverage of base executors

* Add FT for electron error reporting

* Improve test coverage for local and dask executors

* Remove DispatchInfo

This has been unused for a while.

* Don't show stdout in Electron ErrorCard

The ErrorCard shouldn't appear for normally running electrons.

* Change log priority for some routine logging statements

* Update changelog

* Update changelog

* Add boilerplate to alembic migration script

* Revert "Remove DispatchInfo"

This reverts commit ec77cc22fad97026080ce95b7b906fffddcad85e.

* Improve coverage for upsert

* Improve coverage for upsert

* Fix changelog after merge

* Refactored local and dask to deduplicate wrapper code

local and dask wrappers are now stored in a separate module under
covalent/executor/utils