Fix inheritance of sublattice metadata
Parent lattice's metadata are copied to a sublattice if the client did
not specify the sublattice's corresponding metadata. For example

If Lattice executor = "awslambda" and Sublattice executor = <unset>,
the sublattice will get executor="awslambda".

But if Lattice executor = "awslambda" and Sublattice executor = "awsbatch",
the sublattice executor will be "awsbatch".

The one exception to this rule is "workflow_executor", which must be
set and is currently not propagated.