Initial port of SDK changes for task packing
Each transport graph node acquires a new property `task_group_id`
which defaults to the `node_id`. Auto-generated electrons for
collection and attribute access are "packed" with and executed in the
same compute job as the main electron; an exception is made for
sublattice electrons since those aren't "executed" in an executor but
rather dispatched.