Multistage: repack Result.inputs client-side before dispatch
When the Result object is constructed, result.inputs is initialized to
a dictionary
```
{"args": lattice.args, "kwargs": lattice.kwargs}
```
where `lattice.args` and `lattice.kwargs` are collections of
TransportableObjects, and the dictionary is itself pickled to
Covalent's object store. When the UI wants to render the workflow inputs, it needs to unpickle
this dictionary and extract the `object_string` of each
TransportableObject.

With this change, the SDK repacks the inputs into a single
TransportableObject before shipping the dispatch to the server. This
will allow us to serve the workflow inputs' string representation more
efficiently when coupled with the new TransportableObject format.